+++ const/nsq.go
@@ -2,6 +2,8 @@ package _const
 
 //Write by zhangtao<ztao8607@gmail.com> . In 2018/2/5.
 
+import zmodel "github.com/openzipkin/zipkin-go/model"
+
 const (
 	SvcDestroyMsg    = "DestroySvc"
 	SvcMonitorMsg    = "MonitorSvc"
@@ -11,6 +13,7 @@ const (
 )
 
 type DestoryMsg struct {
-	Svcname   string `json:"svcname"`
-	Namespace string `json:"namespace"`
+	Svcname   string             `json:"svcname"`
+	Namespace string             `json:"namespace"`
+	Span      zmodel.SpanContext `json:"span"`
 }